materialApoio/Senti-Pred-Remake/requirements.txt
pandas==2.2.3
numpy==1.26.4
torch==2.7.1
transformers==4.49.0
datasets==3.2.0
scikit-learn==1.6.1
matplotlib==3.10.0
seaborn==0.13.2
mlflow==2.20.2
dagshub==0.3.41
python-dotenv==1.0.1
accelerate==1.3.0
